requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
pydantic[email]==2.5.0
requests==2.31.0
scikit-learn==1.3.2
numpy==1.24.3
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.1
python-multipart==0.0.6
bcrypt==4.0.1
passlib==1.7.4
python-jose[cryptography]==3.3.0
h3>=4.0.0
geopy==2.4.0
aiofiles==23.2.1
stripe==7.9.0
pillow==10.1.0
openai==1.3.0
filetype==1.2.0
psutil==5.9.6
firebase-admin==6.4.0
google-cloud-firestore==2.14.0
appwrite==4.0.0
python-dotenv==1.0.0